+++ unreal/AegisWar/Source/AegisWar/Public/WarPlayerController.h
@@ -24,6 +24,8 @@ public:
     void ToggleQuestLog();
     void ToggleWorldEditor();
     UFUNCTION(Server, Reliable) void ServerEditWorldObject(FName Id, FTransform Transform, bool bObjectHidden, int32 ExpectedRevision);
+    UFUNCTION(Server, Reliable) void ServerCreateWorldObject(FName TemplateId, FTransform Transform, int32 ExpectedRevision);
+    UFUNCTION(Client, Reliable) void ClientWorldObjectCreated(FName Id);
     UFUNCTION(Server, Reliable) void ServerWorldEditHistory(bool bRedo, int32 ExpectedRevision);
     UFUNCTION(Server, Reliable) void ServerWorldEditDraft(bool bLoad, int32 ExpectedRevision);
     UFUNCTION(Client, Reliable) void ClientWorldEditResult(const FString& Message);