+++ unreal/AegisWar/Source/AegisWar/Public/WarWorldEditSubsystem.h
@@ -14,6 +14,7 @@ class AEGISWAR_API UWarWorldEditSubsystem : public UWorldSubsystem
 public:
     bool CanUse(const APlayerController* Controller) const;
     bool Open(APlayerController* Controller, FString& Error);
+    bool Create(APlayerController* Controller, FName TemplateId, const FTransform& Transform, int32 Revision, FName& CreatedId, FString& Error);
     bool Edit(APlayerController* Controller, FName Id, const FTransform& Transform, bool bHidden, int32 Revision, FString& Error);
     bool Undo(APlayerController* Controller, bool bRedo, int32 Revision, FString& Error);
     bool SaveDraft(APlayerController* Controller, int32 Revision, FString& Error);
@@ -23,7 +24,20 @@ public:
     FString GetDraftLocation() const;
 private:
     bool Ready(APlayerController* Controller, FString& Error);
+    bool ApplyHistory(FWarWorldEditHistory Next, FString& Error);
     void ApplyActors();
+    struct FCollisionTemplate
+    {
+        FTransform Transform;
+        FVector Extent;
+        FName Profile;
+    };
+    struct FModelTemplate
+    {
+        TWeakObjectPtr<class UStaticMesh> Mesh;
+        TArray<FCollisionTemplate> Collision;
+    };
+    TMap<FName, FModelTemplate> Templates;
     FWarWorldEditHistory History;
     TMap<FName, TWeakObjectPtr<AActor>> Actors;
     FString LastDiskContents;